DDPG and PPO working implementations

+++ src/model/ddpg/buffer_replay.py
@@ -5,9 +5,11 @@ import torch
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
+
 class ReplayBuffer():
     experience = namedtuple('Experience', field_names=['state', 'action', 'reward', 'next_state', 'done'])
     experience.__qualname__ = 'ReplayBuffer.experience' # necessary for pickling
+
     def __init__(self, action_size, buffer_size, batch_size, seed):
         """Initialises the Replay Buffer
 
@@ -22,7 +24,7 @@ class ReplayBuffer():
         self.batch_size = batch_size
         self.seed = random.seed(seed)
         self.memory = deque(maxlen=buffer_size)
-    
+
     def add(self, state, action, reward, next_state, done):
         """adds element to the buffer
 
@@ -30,7 +32,7 @@ class ReplayBuffer():
             state ([float]): state vector
             action ([float]): action indice
             reward ([float]): reward
-            next_state ([float]): next state vector 
+            next_state ([float]): next state vector
             done ([boolean]): boolean indicating the end of the episode
         """
         e = self.experience(state, action, reward, next_state, done)
@@ -43,20 +45,19 @@ class ReplayBuffer():
             tuple: tuple of experience arrays.
         """
         batch_values = random.sample(self.memory, k=self.batch_size)
+
         state = torch.from_numpy(np.vstack([e.state for e in batch_values])).float().to(device)
-        action = torch.from_numpy(np.vstack([e.action for e in batch_values])).long().to(device)
+        action = torch.from_numpy(np.vstack([e.action for e in batch_values])).float().to(device)
         reward = torch.from_numpy(np.vstack([e.reward for e in batch_values])).float().to(device)
         next_state = torch.from_numpy(np.vstack([e.next_state for e in batch_values])).float().to(device)
         done = torch.from_numpy(np.vstack([e.done for e in batch_values]).astype(np.uint8)).float().to(device)
         return (state, action, reward, next_state, done)
-    
+
     def __len__(self):
         return len(self.memory)
 
-
-
 class PrioritizedReplayBuffer(ReplayBuffer):
-    
+
     def __init__(self, action_size, buffer_size, batch_size, seed):
         """Initialises the Priotity Replay Buffer
 
@@ -69,8 +70,8 @@ class PrioritizedReplayBuffer(ReplayBuffer):
         super().__init__(action_size, buffer_size, batch_size, seed)
         self.priorities = deque(maxlen=buffer_size)
         self.sampled_idx = None
-        
-    
+
+
     def add(self, state, action, reward, next_state, done):
         """adds element to the buffer and initialises it's priority to the maximum available
 
@@ -78,7 +79,7 @@ class PrioritizedReplayBuffer(ReplayBuffer):
             state ([float]): state vector
             action ([float]): action indice
             reward ([float]): reward
-            next_state ([float]): next state vector 
+            next_state ([float]): next state vector
             done ([boolean]): boolean indicating the end of the episode
         """
         super().add(state, action, reward, next_state, done)
@@ -86,16 +87,16 @@ class PrioritizedReplayBuffer(ReplayBuffer):
             self.priorities.append(max(self.priorities))
         else:
             self.priorities.append(1.0)
-    
+
     def sample(self, a=0.9):
         """samples a batch of experience arrays from the buffer according to their priority
-        
+
         Args:
             state ([float]): state vector
         Returns:
             tuple: tuple of experience arrays.
         """
-        p = np.power(self.priorities, a)/ np.power(self.priorities, a).sum()
+        p = np.power(self.priorities, a) / np.power(self.priorities, a).sum()
         sample_idx = random.choices(np.arange(len(self.priorities)), k=self.batch_size, weights=p)
         state = torch.from_numpy(np.vstack([self.memory[i].state for i in sample_idx])).float().to(device)
         action = torch.from_numpy(np.vstack([self.memory[i].action for i in sample_idx])).float().to(device)
@@ -104,7 +105,7 @@ class PrioritizedReplayBuffer(ReplayBuffer):
         done = torch.from_numpy(np.vstack([self.memory[i].done for i in sample_idx]).astype(np.uint8)).float().to(device)
         self.sample_idx = sample_idx
         return (state, action, reward, next_state, done, p[sample_idx])
-    
+
     def update(self, priority_vals):
         for idx, p in zip(self.sample_idx, priority_vals):
             self.priorities[idx] = p
